+++ rethinkdb-net-test/Integration/SingleObjectTests.cs
@@ -5,6 +5,7 @@ using System.Net;
 using System.Linq;
 using System.Threading.Tasks;
 using System.Linq.Expressions;
+using System.Collections.Generic;
 
 namespace RethinkDb.Test.Integration
 {
@@ -34,6 +35,12 @@ namespace RethinkDb.Test.Integration
             {
                 Name = "Jim Brown",
                 Children = new TestObject[] {
+                    new TestObject() { Name = "Scan" }
+                },
+                ChildrenList = new List<TestObject> {
+                    new TestObject() { Name = "Scan" }
+                },
+                ChildrenIList = new List<TestObject> {
                     new TestObject() { Name = "Scan" }
                 },
                 SomeNumber = 1234